+++ CHANGELOG.md
@@ -9,11 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Remove elements with `opacity="0"`.
 - Transfer the group `id` attribute to the child when group has only one child.
 - `symbol` element support.
+- `parse_tree_from_str`.
 
 ### Changed
 - Rename `Tree::node_by_svg_id` to `Tree::node_by_id`.
 - Use `rctree::Node<NodeKind>` instead of `rctree::Node<Box<NodeKind>>`.
 - `view` element is out of scope now.
+- `FileReadError` replaced with `Error`.
+- `parse_tree_from_data` accepts `&[u8]` and not `&str` now.
 
 ### Removed
 - `NodeExt::kind`. Use `Node::borrow` instead.